erste gute Version zum sharen. ich brauche aber noch url und bild.

+++ YourGoalsKit/ShareExtensionTasksProvider.swift
@@ -8,6 +8,11 @@
 
 import Foundation
 
+enum ShareExtensionError : Error {
+    case taskNameNeeded
+}
+
+
 /// class for saving new tasks from the share extension and consume then
 public class ShareExtensionTasksProvider {
     
@@ -21,17 +26,23 @@ public class ShareExtensionTasksProvider {
         self.manager = manager
     }
     
-    /// create a new task from a shared extenison
+    /// save a new task request from a shared extenison in a buffer storage
     ///
     /// - Parameters:
     ///   - name: name of the new task
     ///   - url: an optional url
     ///   - image: an optional image
     /// - Throws: core data exception
-    public func saveNewTaskFromExtension(name: String, url: URL?, image:UIImage?) throws {
+    public func saveNewTaskFromExtension(name: String, url: String?, image:UIImage?) throws {
         let shareNewTask = manager.shareNewTaskStore.createPersistentObject()
+        
+        guard !name.isEmpty else {
+            throw ShareExtensionError.taskNameNeeded
+        }
+        
+        shareNewTask.taskname = name
         shareNewTask.image = image?.jpegData(compressionQuality: 0.7)
-        shareNewTask.url = url?.absoluteString
+        shareNewTask.url = url
         try manager.saveContext()
     }
     
@@ -39,10 +50,14 @@ public class ShareExtensionTasksProvider {
     ///
     /// - Parameter consume: function to process the shared task
     /// - Throws: exception
-    public func consumeNewTasksFromExtension(consume: (ShareNewTask) -> () ) throws {
+    public func consumeNewTasksFromExtension(consume: (ShareNewTask) throws -> () ) throws {
         let newTasks = try self.manager.shareNewTaskStore.fetchAllEntries()
         for task in newTasks {
-            consume(task)
+            guard task.taskname != nil && !task.taskname!.isEmpty else {
+                continue
+            }
+            
+            try consume(task)
             self.manager.context.delete(task)
         }
